backend/requirements.txt
fastapi
uvicorn
python-multipart
torch>=2.0.0
transformers>=4.40.0
flash-attn>=2.5.0
qwen-vl-utils
pillow
httpx
python-multipart

